save current place just for sake of it
